# Remove RC4 from allowed SSL ciphers