@tiwillia 2016-01-19
[test][extended:node]

@Miciah PTAL
@thrasher-redhat fyi