@sferich888 2016-01-19
@tiwillia so were going to limit the node-web-proxy to only AES256-GCM-SHA384 ciphers, and not allow or fix the issue define in https://bugzilla.redhat.com/show_bug.cgi?id=1299014 to make the cipher that this service uses by default, configurable?