@rjhowe 2016-01-19
The ciphers need to be all listed if we are defining these in the code. 

After the change. 
~~~
[root@node1 ~]# nmap --script ssl-cert,ssl-enum-ciphers -p 8443 localhost

Starting Nmap 5.51 ( http://nmap.org ) at 2016-01-19 18:01 EST
Nmap scan report for localhost (127.0.0.1)
Host is up (0.00014s latency).
Other addresses for localhost (not scanned): 127.0.0.1
PORT     STATE SERVICE
8443/tcp open  https-alt
| ssl-cert: Subject: commonName=*.example.com/organizationName=OpenShift Enterprise default/stateOrProvinceName=SomeState/countryName=XX
| Issuer: commonName=*.example.com/organizationName=OpenShift Enterprise default/stateOrProvinceName=SomeState/countryName=XX
| Public Key type: rsa
| Public Key bits: 2048
| Not valid before: 2015-10-15 19:18:36
| Not valid after:  2025-10-12 19:18:36
| MD5:   0a64 16e5 29e1 af48 6c02 c78c b155 1d1a
|_SHA-1: 795b a647 13b8 9367 7c00 bc74 b769 d3af 894c 5e92
| ssl-enum-ciphers: 
|   TLSv1.2
|     Ciphers (1)
|       TLS_RSA_WITH_AES_256_GCM_SHA384
|     Compressors (1)
|_      uncompressed

Nmap done: 1 IP address (1 host up) scanned in 1.70 seconds
~~~

Before the Change 
~~~
[root@node1 ~]# nmap --script ssl-cert,ssl-enum-ciphers -p 8443 localhost

Starting Nmap 5.51 ( http://nmap.org ) at 2016-01-19 17:59 EST
Nmap scan report for localhost (127.0.0.1)
Host is up (0.00017s latency).
Other addresses for localhost (not scanned): 127.0.0.1
PORT     STATE SERVICE
8443/tcp open  https-alt
| ssl-cert: Subject: commonName=*.example.com/organizationName=OpenShift Enterprise default/stateOrProvinceName=SomeState/countryName=XX
| Issuer: commonName=*.example.com/organizationName=OpenShift Enterprise default/stateOrProvinceName=SomeState/countryName=XX
| Public Key type: rsa
| Public Key bits: 2048
| Not valid before: 2015-10-15 19:18:36
| Not valid after:  2025-10-12 19:18:36
| MD5:   0a64 16e5 29e1 af48 6c02 c78c b155 1d1a
|_SHA-1: 795b a647 13b8 9367 7c00 bc74 b769 d3af 894c 5e92
| ssl-enum-ciphers: 
|   TLSv1.0
|     Ciphers (6)
|       TLS_RSA_WITH_3DES_EDE_CBC_SHA
|       TLS_RSA_WITH_AES_128_CBC_SHA
|       TLS_RSA_WITH_AES_256_CBC_SHA
|       TLS_RSA_WITH_CAMELLIA_128_CBC_SHA
|       TLS_RSA_WITH_CAMELLIA_256_CBC_SHA
|       TLS_RSA_WITH_RC4_128_SHA
|     Compressors (1)
|       uncompressed
|   TLSv1.1
|     Ciphers (6)
|       TLS_RSA_WITH_3DES_EDE_CBC_SHA
|       TLS_RSA_WITH_AES_128_CBC_SHA
|       TLS_RSA_WITH_AES_256_CBC_SHA
|       TLS_RSA_WITH_CAMELLIA_128_CBC_SHA
|       TLS_RSA_WITH_CAMELLIA_256_CBC_SHA
|       TLS_RSA_WITH_RC4_128_SHA
|     Compressors (1)
|       uncompressed
|   TLSv1.2
|     Ciphers (10)
|       TLS_RSA_WITH_3DES_EDE_CBC_SHA
|       TLS_RSA_WITH_AES_128_CBC_SHA
|       TLS_RSA_WITH_AES_128_CBC_SHA256
|       TLS_RSA_WITH_AES_128_GCM_SHA256
|       TLS_RSA_WITH_AES_256_CBC_SHA
|       TLS_RSA_WITH_AES_256_CBC_SHA256
|       TLS_RSA_WITH_AES_256_GCM_SHA384
|       TLS_RSA_WITH_CAMELLIA_128_CBC_SHA
|       TLS_RSA_WITH_CAMELLIA_256_CBC_SHA
|       TLS_RSA_WITH_RC4_128_SHA
|     Compressors (1)
|_      uncompressed

Nmap done: 1 IP address (1 host up) scanned in 1.67 seconds
~~~