@tiwillia 2016-01-27
@Miciah @sferich888 Whats the consensus on this? Is it ready for merge?

Do you think we should be marking this as a configuration file in the rpm spec from here-on-out?