@Miciah 2016-01-27
If we add `%config` this release, we will need to make sure `web-proxy-config.json` still gets updated, or that `ciphers` gets a reasonable default value even if it isn't set in `web-proxy-config.json`.

My preference is that we specify a default value in the assignment to `ssl_opts.ciphers` in `http-utils.js`.  It would be nice also to specify `%config web-proxy-config.json` in the spec file.