@tiwillia 2016-01-27
Specifying a default for just this parameter is inconsistent, but I agree that it is the only option if we are going to mark this as a configuration file. I'll work on that, thanks.