@tiwillia 2016-01-27
@Miciah could you give my latest changes a look when you have a chance?