@Miciah 2016-01-27
Looks good! I can tag it for merge after the tests run.