@tiwillia 2016-01-28
@Miciah Bug opened for the configuration file issue: https://bugzilla.redhat.com/show_bug.cgi?id=1302787